Add DIT participants to interaction admin as a read-only section
This adds a DIT participants section to the interaction form in the admin site.

This also adds help text for the deprecated `dit_adviser` and `dit_team` fields to say that they are deprecated. (An existing migration was edited to prevent unnecessary foreign key constraint drop and recreation operations when the schema hasn't been modified.)