Revert redis-py and Kombu updates
Celery 4.2.2 was released but they restricted redis to <3 and Kombu to <4.4 in it (even though Celery 4.2.1 has been out for some time and did not have a redis<3 restriction).

Downgrading redis-py to 2.x now seems risky so this reverts things to how they were in the last release.

When Celery 4.3 is released, we can look at updating these again.