Add export, investment and other service contexts
This adds additional service contexts to support a new question being added in the front end when creating company interactions (about whether the interaction is export-related, investment-related or something else).

The `interaction` and `service_delivery` contexts are also deprecated and will be removed once the front end is using the new contexts.