Update Celery to version 4.3
This includes changing a test which was passing a function to a Celery task (which was invalid as a function cannot be serialised as JSON).

The upgrade also allowed us to greatly simplify our rediss:// handling in our Celery configuration.

Some other change also allowed us to simplify the rediss:// configuration for caching as well.

The Redis configuration changes have been tested in the UAT environment.