Check for encoding errors earlier on when loading marketing opt-outs
This changes the logic in the form so that Unicode decoding errors after the first line in the CSV file are always picked up in the form's clean() method.

This means that any such file is rejected before an attempt to process it is made and that the logic in save() can be simplified.