Exclude matches with already allocated duns_numbers in dnb cleanse cmd
Before: the command did not take into account matches that had a
duns number which had already been used in another data hub record.

After: the command is now excluding duns numbers that have already
been allocated to data hub Company records.

This is because there's a unique constraint on Company.duns_number.

It was not a massive problem as the logic would have handled the error
but this way is more elegant as those cases are ignored.