Add theme field to Interaction model
This field will record the answer to a new question that is being added to the interaction flow.

The main purpose of collecting this value is to filter the service list displayed to the user in the interaction form (rather than collecting the data for analysis or similar purposes).

In particular, the field needs to be saved so that the correct service list is displayed to the user when editing an interaction.

Eventually, we will probably be able to derive this information from the service instead of this field. However, that is a long-term view.

rebase