Refactor stream handling in BaseCSVImportForm
This stops the overwriting of the `csv_file` field with a `csv.DictReader` so that the raw file can be read later on (in order to store its contents).

It also adds a couple of methods to `BaseCSVImportForm` for opening the CSV file as a text stream or as a `csv.DictReader`.