Correct various search sort-by values
In #1643 the sort by options for some search apps were changed, when we should've used remappings instead to keep API compatibility. This corrects that.

However, none of these particular sort-by values appear to be in use. We should, in fact, deprecate them and add logging to make sure they aren't being used. I have left that for a follow-up task.