Add saving-related views to import interactions admin site tool
This adds a button to the preview page import interactions tool that performs the actual saving of loaded rows, and a confirmation page that is displayed afterwards.

The standard Django pattern of redirecting to another page when a form is submitted and changes committed is used. (This pattern limits the possibility of form resubmissions on page refresh.)

All state is kept in the cache for consistency.