Use index to speed up looking up advisers in import interactions tool
This changes the adviser look-up in the import interactions tool to use the `||` SQL operator instead of the `concat_ws` function and creates an index that can be used to speed up the query.

The reason for switching away from `concat_ws` is that it is not possible to use it in indexes. (That is not allowed because for some inputs (e.g. numbers and timestamps) the output can depend on the environment.)