Disable atomic requests for /whoami/
As the `/whoami/` endpoint does not directly modify the database, this opts it out of atomic requests so that it does not cause a transaction to be used for the lifetime of the request.

This should also ensure that when OAuth2 token introspection causes an access token to be inserted into the database (for requests to this endpoint), it is visible to other requests as soon as possible.