Make OMIS nested views compatible with DRF schema introspection
This refactors the views slightly so that `Http404`s are not raised when building serializer contexts if the order referenced in the URL path does not exist (which breaks DRF schema introspection).

Instead, this is checked as part of the request dispatch process (which seems like a more logical place to do this kind of check).