Enable the Django Rest Framework built-in documentation feature
This enables the Django Rest Framework built-in documentation feature at the URL path `/docs`.

This is currently only enabled if the `ENABLE_API_DOCUMENTATION` environment variable is set to `True` as the documentation is not fully functional as yet.

A user must log into Django admin prior to accessing the documentation.

The DRF 3.9 will move more to OpenAPI-based schema generation and these changes will help us to start using that when we upgrade.