Add endpoint to add a company to a user's personal list
This adds an endpoint at `PUT /v4/user/company-list/<company ID>` which adds a company to the authenticated user's personal list of companies.

A generic view has not been used in this case as the behaviour required for these endpoints is somewhat different.

At the moment, the serializer does not do much, but validation that added companies aren't archived is likely to be added.