Update Django Rest Framework and django-filter
This updates Django Rest Framework to version 3.10.1 and django-filter to version 2.2.0.

This keeps DRF using Core API schema features. Switching to OpenAPI will be tackled separately.

There was a slight change to the order DRF saves fields. Previously, it saved to-one and to-many fields in a somewhat arbitrary order. Now, it consistently saves to-many fields last. One of the investment project GVA tests was failing after updating DRF as there was an assumption that the business activities to-many field would be saved before the model object. This has been fixed by adding an additional signal receiver that monitors changes to the `business_activities` field specifically.