Add Gunicorn StatsD configuration
This enables StatsD instrumentation for Gunicorn.

Both the Cloud Foundry instance index and GUID are included in the metric prefix so that metrics for different instances can never overwrite each other. We should be able to aggregate them in Grafana as needed.

A new environment variable has been added to control whether this is enabled. This is so we can opt in to it as needed, and also because Gunicorn logs warnings if it can't connect to the StatsD server.