Tidy up investment project search model slightly
Originally, we left dynamic field mapping on in Elasticsearch. This resulted in a few fields being unintentionally added to the investment project mapping type.

Later on, we turned off dynamic field mapping and explicitly added the fields that we had accidentally indexed to the mapping type.

This makes an effort to tidy things up slightly by turning indexing off on those fields (as they are not used in searches). The field definitions have also been simplified as a result.

The fields have still been kept in the mapping type for now to avoid changing API responses.

In one case, one of the fields had been added as a sorting option. However, it was non-functional as it is not possible to sort on `text` fields. Hence, I've removed that field from the allowed sorting options.