Use CompanyListItem.list in all legacy company list endpoints
Now that `CompanyListItem.list` has been populated for all existing list items, and is non-nullable, this makes the existing legacy company list endpoints operate on the default company list for each user.

(This means querying based on `CompanyListItem.list.adviser` and `CompanyListItem.list.is_legacy_list` instead of `CompanyListItem.adviser`.)

This is so we can work on the functionality that will allow users to have multiple company lists while having the existing functionality continue to work unaffected. It will also allow us to seamlessly put multiple list functionality live when we are ready.