Fix random failure of delete_old_records test
This corrects one of the `delete_old_records` OMIS order test parametrisations where a value for one of the datetime fields was not being explicitly specified.

The value was intended to be `None`, but it was picking up a random datetime which was sometimes not before the cut-off date for deleting old records.