Deprecate unused contact search sorting options
This deprecates various sorting options for contact search that aren't being used by the front end.

This is so that we can simplify the code base and make future refactoring easier. It will also allow us to make the contact search model leaner.

Many of these sorting options were also untested.

Logging has been added to make sure we don't break things inadvertently.

The approach used is the same as in #1450.