Deprecate unused investment project search sorting options
This deprecates various sorting options for investment project search that aren't being used by the front end.

This is so that we can simplify the code base and make future refactoring easier. It will also allow us to make the investment project search model leaner.

Many of these sorting options were also untested.

Logging has been added to make sure we don't break things inadvertently.

The approach used is the same as in #1450.