Use consistent sub-package name for tests
Most commonly, `test` has been used for the name of the sub-package containing tests for each Django and search app.

However, over time, a few uses of `tests` (with 's') have crept in. Since `test` was already established and is much more common, this renames the ``tests`` packages to ``test``.

(There were about 39 `test` directories and 9 `tests` directories prior to this change.)