Add item count to company list responses
This adds an `item_count` field to the responses of the following company list endpoints:

- `GET /v4/company-list`
- `PATCH /v4/company-list/<id>`

This field contains the count of items contained in the company list.

(It does not get added to the POST response, because the created model object doesn't have the annotation.)

This is being added for the 'Edit lists' page in the front end where an item count is displayed for each list.