Split removal and deprecation sections in changelog
This updates the towncrier configuration so that `removal` and `deprecation` are separate news fragment types (and hence separate sections in the changelog).

This is to improve clarity (as opposed to mixing the two things under one heading).

This was discussed some time ago, but we didn't get round to making the change until now.