Convert changelog to Markdown
This converts the changelog (including news fragments) to use Markdown instead of reStructuredText.

We decided to make this change as we find Markdown easier to use and switching to it should lead to fewer formatting mistakes.

All future news fragments will need to be formatted using Markdown and use the `.md` extension. (This includes any in any currently open PRs.)