Add script to automate creation of changelog
This adds a script that automates the tasks associated with creating a changelog.

Python was used for the script because:

- we're all familiar with it
- it's portable
- it handles arguments well

Most of the tasks associated with creating a changelog can now be performed by running `scripts/create_changelog.py major|minor|patch`.

The script will increment the version automatically, create a branch with the updated changelog, push the new branch, and open your web browser ready to create a PR.

The release instructions have been updated to reflect these changes.