Fix intermittently failing company autocomplete tests
This fixes two intermittently failing company autocomplete search tests.

The tests assumed that the two companies would be returned in a particular order, but this was not guaranteed because the two documents had the same term frequency for the search term (our code does some deduplication when building the suggestion terms during indexing).

This adds the search term to another company field for one of the objects used in the test, in order to increase the term frequency for that object to make sure it is returned first.