Apply schema migrations for the mi database automatically on deployment
Prior to this change, migrations for the `mi` database had to be applied manually in deployed environments, and we would often forget to do this.

This change modifies the script used for the Procfile web process so that migrations for that database are applied automatically.

Because we don't have an MI database in some environments, this step can be skipped by setting the `SKIP_MI_DATABASE_MIGRATIONS` environment variable to a non-empty value.