Fix leaking SimpleModel instance in search tests
There was a problem when running tests using `--reuse-db` where a couple of tests would fail because there was a SimpleModel object left over from another test.

This fixes the leaking SimpleModel instance which was tracked down to a test that was using a thread pool task. Because the test was using another thread, it has to clean up manually.

This changes the test in question to use a transaction and a rollback to bring it line with what the tests usually do.