Rename the setup_es fixture to es_with_signals
This renames the `setup_es` pytest fixture to `es_with_signals` to be more explicit about the fact that it connects the search signal receivers.

This is in advance of a separate fixture being added that does bulk syncing to Elasticsearch rather than relying on signal receivers syncing objects individually (which is slow, and means we are testing the signal receivers in every test).