Split es_with_signals fixture
This splits off an `es` fixture from the `es_with_signals` fixture. This fixture enables Elasticsearch use during a test without connecting signal receivers.

This is in advance of a separate fixture being added that does bulk syncing to Elasticsearch rather than relying on signal receivers syncing objects individually (which is slow, and means we are testing the signal receivers in every test).