Migrate Elasticsearch mapping types automatically on deployment
This schedules the `migrate_es` management command to run automatically on deployment.

Previously, we had to run this command manually after deploying. However, this relied on:

- the person deploying knowing it needed to be done; and
- remembering to do it in every environment

There is a slight bit of complexity in automating this in that there is a chance that, during the rolling deployment, an old Celery instance could pick up the tasks scheduled by `./manage.py migrate_es`. However, there is already logic to handle this in the `complete_model_migration` task and it appears to work well from testing in the `uat` environment.

Nonetheless, the `SKIP_ES_MAPPING_MIGRATIONS` environment variable can be set to skip running `migrate_es`.