Add an API endpoint to export Teams
This is similar to Advisors #2187 and other dataset endpoints,
but since teams don't have a creation date we're ordering the
table by ID only.

This has a small chance of potentially breaking pagination, but
(based on a discussion with @reupen) adding a creation date to the
model is out of scope for this and the table data doesn't change
that often so the risk is low.

Since Teams are pre-populated by a session-wide fixture the test
endpoint returns a lot of existing results. So instead of creating
new ones we check serialization for one of the existing teams and
test that items are sorted by ID.