Handle both strings and UUIDs correctly in InvestmentProject property
The `InvestmentProject.level_of_involvement_simplified` property had an assumption in the logic that `level_of_involvement_id` would be a UUID instance. However, it is valid to assign `level_of_involvement_id` a UUID in string form and this can happen when constants from our code base are used.

This corrects the logic of the property so that both `UUID` objects and strings are handled correctly.

A new utility function, `force_uuid()` was added to help with this.