Add a dataset base view test class and shared fixtures
We have some largely identical tests that are repeated for each view.
They tests authentication and basic responses and the only differences
are the view url and the object factory used to set them up, which can
be easily configured with class attributes.

This moves common tests to a base class in dataset.core.test. Since
the tests depend on pytest fixtures we need to move them to a shared
conftest to avoid unused imports in other test modules.

The base class could be in the conftest as well, but I think it makes
more sense alongside the base view itself.