Add remove company account manager endpoint
This adds a new endpoint, `POST /v4/company/<ID>/remove-account-manager`, that removes the assigned One List tier and account manager for a company on One List tier 'Tier D - Interaction Trade Adviser Accounts'.

If the company is on a One List tier other than 'Tier D - Interaction Trade Adviser Accounts', the operation is not allowed.

The `company.change_company` and `company.change_regional_account_manager` permissions are required to use this endpoint.

A similar pattern to the self-assign account manager view and other action views in the code base was followed.