Fix intermittently failing event search sorting test
This fixes an intermittently failing test for sorting event search results by `modified_on:desc`.

Locally this was failing about 6% of the time for me. It was intermittently failing because it was relying on the system clock advancing by more than a millisecond (the resolution used by Elasticsearch `Date` fields) between two save operations.

This changes the test to use `freeze_time()` instead of relying on the system clock. (An extra `save()` call was also removed, as it's not relevant for the test.)