Stop pip cache growing unbounded on CircleCI
The pip cache on CircleCI had grown to 1.1 GB, because packages were added to it but never removed (e.g. when updated).

This changes the configuration to only restore a cache that matches the the specific set of requirements we currently have (whereas before it would fall back to the most recent cache if an exact match for the requirements wasn't found).

This stops the cache growing unbounded as a result, and gets its size down to about 40 MB.

Most builds should be about 15 seconds faster (builds where dependencies are changed will take about the same time as before).

The cache key was also changed so that it includes the Docker image name (so that the key includes the Python version).