Optimise entity search tests by syncing only the app being tested
This adds a simple optimisation to the entity search view tests so that only the search app being tested is synced to Elasticsearch.

This is done by making the `es_with_collector` fixture only sync the apps set in the `es_collector_apps` marker (if it is set).

(I went for this approach as it is easy to set a marker on a whole module.)

(This optimisation is possible as each entity search view only searches the index for a particular search app.)

This reduces the running time of `pytest -n2 --reuse-db datahub/search/` by around 13% for me (locally).