Make shared basic search tests generic
This deals with (mostly) two TODOs about making some shared tests for basic (global) search generic.

(Previously, these tests were relying on specific search apps when the tests were actually testing generic logic. Using an existing search app gave the impression that the tests were specific to that search app which was not the case.)

One test for permission handling was not tackled as it was a bit more involved. It has been left as a future exercise.