Squash all investment project migrations
This squashes all investment migrations. It includes some manual optimisation, corrections and clean-up of the replacement migration.

Once this has been deployed to all environments, the old migrations can be removed and the squashed migration can be turned into a normal migration.

See https://docs.djangoproject.com/en/2.2/topics/migrations/#squashing-migrations for more information on squashing migrations.

Note that there was a circular dependency between the `investment` and `core` migrations that had to be broken. The `core` migrations are now irrelevant so the dependency was simply removed.