Optimise cleanup management command tests
This optimises tests for the `cleanup` app so that they sync objects to Elasticsearch in bulk during set-up.

These tests need to use the normal Elasticsearch signals when the command is run, so a new pytest fixture was added that can be used to temporarily activate the Elasticsearch collector (so that it can be used for test set-up only in this case).

Locally, this reduces the time to run `pytest --reuse-db datahub/cleanup/` from 49 seconds to 27 seconds.