Fix occasionally failing basic search test
This fixes an occasionally failing test for basic search pagination.

The test was occasionally failing because it was sorting the IDs of the created `SimpleModel` objects as numbers, when Elasticsearch was sorting them as strings (as `id` is a `Keyword` field on the search model).

Hence, the test was sometimes failing depending on what the `id` values of the model objects created in the test were.