Fix error when merging two companies on same company list
This fixes an unhandled error when merging two companies that are both on the same company list.

The intention for this case was that any references to the archived company on lists would be updated to refer to retained company.

However, if both companies were on the same company list, the previous logic was trying to create a duplicate company list item which resulted in an `IntegrityError` being raised.

This has been corrected so that in this particular case, the archived company is removed from the list instead.

This wasn't that easy to fix since the logic required for this is different from the general case – I've tried to incorporate the required logic in a non-obtrusive way.