Add a script to publish a release on GitHub
This adds a new script, `script/publish_release.py`, that publishes a release on GitHub once it has been merged to master.

The script works without checking out a different branch, as it does not need to commit anything.

A GitHub access token with the `public_repo` scope is required to run the script.