Add interaction DIT participants to company list item response
This adds a list of DIT participants to the latest interaction of each company returned by `GET /v4/company-list/<pk>/item`.

As a reminder, the retrieval of this data is done in the main query to avoid doing individual queries for each company list item. This was tricky in this case, as each interaction can have multiple participants. I came up with a somewhat creative solution of combining `ArrayAgg` and `JSONBBuildObject` to do this.

From checking one of the current worst cases in production, the execution time of the query according to `EXPLAIN ANALYSE` goes up from around 4ms to 9ms (which is still fast).