Fix missed root news fragments in release scripts
Just spotted that the `list_news_fragments()` function missed news fragments in the root of `changelog/` as it was not passing `recursive=True` to `iglob()`.

This would've caused `scripts/prepare_release.py` to fail if there were only root news fragments, and it would've also failed to return an error if a misnamed root news fragment was left behind.

The fix is more complicated than just passing `recursive=True`, as this would also cause `changelog/` sub-directories to be included.

Hence, the function has been rewritten using a different approach, and a test has been added.