Only ignore spaces in company search postcode filter for full postcodes
This changes how the `uk_postcode` filter operates in company search so that e.g. `AB11` and `AB1 1` are distinct searches.

Previously, both searches were equivalent and they both matched the postcodes `AB11 9AA` and `AB1 1AA`. Now, they're treated as distinct searches  so that`AB11` matches `AB11 9AA` and not `AB1 1AA`  (and `AB1 1` matches `AB1 1AA`  and not `AB11 9AA`).

Spaces are still ignored for full postcodes, so that `AB1 1AA` still matches `AB11AA` (and vice versa).

This was done by removing the `space_remover` filter from the `postcode_search_analyzer` analyser and adding a `pattern_replace` filter to both the `postcode_analyzer` and `postcode_search_analyzer` analysers that adds a space in the correct position to full postcodes that are missing one.