Add interaction field to CompanyReferral model
This adds an `interaction` foreign key to the `CompanyReferral` model. This will be used for referrals in the `complete` status.

A fixture representing a complete referral was also added.

The configuration for the `delete_old_records` cleanup management command was updated so that interactions are not deleted if they have any related company referrals. (The general approach is that related records have to be deleted first.)