Add GitHub Actions workflow for publishing releases
This adds a GitHub Actions workflow that automatically publishes a release on GitHub following the merge of a release PR.

The workflow is triggered by the merging of a PR where the base branch is `master` and the head branch starts with `release/`.

The workflow runs the existing `scripts/publish_release.py` script. This script now outputs the URL of the newly published release, instead of opening a web browser window, to better accommodate GitHub Actions. It also now correctly sets the exit code on failure.

The workflow also only installs the few  dependencies required by the script instead of the full set of project dependencies. This was done by using the main `requirements.txt` as a constraints file.

Note: The workflow was tested in the `data-hub-api-actions-test` repo.