Reduce dependency on Elasticsearch mapping type names
This switches from using the mapping type name to the search app name in index and alias names.

This is because mapping type names are deprecated in Elasticsearch and so this will help with upgrading to Elasticsearch 7.

(We use the same values for search app and mapping type names, so there should be no actual change.)

We might be able to move these methods to the search app to make things a bit cleaner, but I've gone for the a less onerous change for now.