Add _document_type field to mapping types
This adds a `_document_type` field to all mappings types as a replacement for the deprecated `_type` field (the mapping type name) which is deprecated in Elasticsearch.

(We still need a type field for the aggregations and permission filtering in global search.)

This change will help with the upgrade to Elasticsearch 7.

Once the new `_document_type` field has been indexed in all environments, queries can be updated to use it.

(I've gone with `_document_type` because `_type` conflicts with the existing field, and `type` or `document_type` could conflict or be confused with a model field.)