Restore ability to merge companies that have export countries
This restores the ability to merge companies (in the admin site) that have export countries.

Note that this does not make any attempt to move the export countries from the source company to the target company. We agreed to just restore the previous behaviour (from when the export countries were stored in many-to-many fields rather than the `CompanyExportCountry` model), where merging was allowed but export countries were left alone.

(This is because merging them would be fairly complex, and can be manually done in the front end if needed.)