Refactor export country history search tests
This refactors the company export country search tests in advance of adding interactions to the search results (without any change to functionality).

This is split from #2632.

The tests are structured in a way that will make it easier to add interactions into the mix.

Additionally. a test for the response body was added.