Add workaround for Django 3 memory leak under gevent
We've been suffering from a problematic memory leak since upgrading to Django 3 which is believed to be caused by https://github.com/django/asgiref/issues/144.

The memory leak is causing Gunicorn worker processes to exhaust available memory and crash and restart.

This adds a workaround which is inactive by default. This is so it can be we can verify that it fixes the problem in e.g. the dev environment.