Add admin site form for adding an adviser using Staff SSO
This adds a new add adviser form that can be used to add an adviser that's been set up in Staff SSO using their SSO email user ID or SSO primary application email address.

This is implemented using a `ModelForm` so it can be used as an `add_form` in a `ModelAdmin`.

A search is first attempted using the SSO email user ID. If no match is found, a search using the email field is then done.

The view will be added separately.