Add admin view for adding an adviser from SSO
This adds a new admin site page for adding an adviser by looking up their details in Staff SSO.

The new page is at `/admin/company/advisor/add-from-sso/`.

Users can be searched for by either their SSO email user ID or any of their email addresses in Staff SSO. Some of their details will also be filled in using the details held in Staff SSO.

This page is currently not linked to; a link will be added on the adviser change list separately.