Allow clients of interactions to set page size
In data-flow, we scrape interactions from this API every morning -
approx 3 mil records as it stands currently. At the moment, this is
paginated in groups of 100 records at a time, which obviously results in
a significant amount of requests. By increasing page size we hope to
make the task generally run faster. At the moment each request seems to
take around 0.1s, so increasing page size by a factor of 10 should be
fine. A factor of 100 should be manageable, though we don't intend to
use that as a client yet.