Refactor caching logic of new introspection authentication class
This breaks out the caching logic of the new introspection authentication class into a separate module.

This is to allow tokens to be added to the cache for development purposes.

The data that is actually stored in the cache was also reduced to what is actually required to make inserting entries for development purposes easier.