Add 'add access token' admin site page
This adds a new admin site page at `/admin/add-access-token`. This can be used to add an access token for development purposes when `STAFF_SSO_USE_NEW_INTROSPECTION_LOGIC` is turned on.

This is to replace doing this via the access token list from Django OAuth Toolkit which does not function when the new introspection logic is in used.

The page can also be disabled using the `ENABLE_ADMIN_ADD_ACCESS_TOKEN_VIEW` setting so we can e.g. turn it off in production.