Enable new SSO token introspection logic by default
This turns on `STAFF_SSO_USE_NEW_INTROSPECTION_LOGIC` by default so the the new SSO token introspection logic is used.

This includes updating tests and `setup-uat.sh` to be compatible with the new logic.

Some tests were passing `Authorization` to the Django Rest Framework test client instead of `HTTP_AUTHORIZATION`. While the former also worked with Django OAuth Toolkit, the latter seems to be the correct argument, and so this has been updated.