Use default mapping type name in Elasticsearch indexes
This updates all Elasticsearch indexes to use the default mapping type name of `_doc`.

This change has been made because custom mapping type names are deprecated. The change will help facilitate the upgrade to Elasticsearch 7.


(The change of mapping type names will cause the migration logic to kick in on deployment, and create new indexes using the new mapping type names. This is because the change of mapping type name causes the calculated mapping hash to change.)