CDS-214 Fixed issue with SQL Schema /ORM metadata mismatch during migration operation
At this point in the migration the ORM metadata and the SQL schema are out of sync because the migration is in the process of updating the schema via the incremental migration scripts (which have been generated from the ORM metadata).
Right now the ORM thinks that the Contact model has an address_area_id field, but the SQL Schema doesn't know about it yet because the migration that adds this field to the SQL schema hasn't yet run.
As part of the delete operation the migration is using the ORM to do a select query on the contact model because the ORM believes there's a foreign key present from the table it's deleting from, but in the SQL Schema the foreign key field doesn't even exist yet. Hence the error:

```
column company_contact.address_area_id does not exist
LINE 1: ...ompany_contact"."id" FROM "company_contact" WHERE "company_c...
```

now that datahub/company/migrations/0112_contact_address_area.py has
been introduced