CDS-311 Add --reuse-db to addopts to reduce upfront timecost of db tests
So, after looking through the profiling output of the database-dependant tests in data-hub-api, it looks like there is an upfront cost to creating the schema of the database that the tests are to be run on of about ~40 seconds. This appears to be run twice, but this should hopefully have been mitigated with the recent removal of the mi database. See the attached image for a screenshot of kcachegrind demonstrating the above.

This marries up with the observations myself and others have made; that there is a large wait time when running even a single database-dependant test in datahub API, so reducing or removing this overhead would seem to be a valuable exercise.

So, django-pytest provides a flag to its testing command to keep the test database after the tests have executed, and if the database exists, reuse it in the current execution: `--reuse-db`
Setting this flag on two concurrent runs caused the test suite execution time of `pytest datahub/dataset/interaction/test/` to drop from 42.85s to 6.12s
This flag has been known about for a while, but people often forget to set it. Fortunately pytest allows you to specify default command line flags in the pytest.ini file by appending them to the `addopts` configuration option. This will allow the database to be reused by default. This can be overwritten by appending `--create-db` to the command line invocation of pytest

If additional migrations have been added since the previous test execution, these seem to be applied even with `--reuse-db` set