DET-51
We are trying to demo getting data from Aventri to the DH front-end. We
aren't able to match up Aventri events to companies yet, so this is an
attempt to just get all Aventri events and print something.

In this commit I'm just printing what the query looks like, and the
response from Aventri.

To run this against the DET-51 branch of the front-end:

* point the backend api to Aventri using the secrets from the Vault in
.env (so far neither dev or staging seem to be returning any Aventri
events)
* generate OAuth token as in readme
* run front-end and go to a companies activities
* should see activities in logs