DET-113 Add `notes` and `theme` to Interaction for Activity Stream
We want to display data from Activity Stream on the Data Hub front end,
so that users have a more complete picture of the data.

We are missing a couple of fields on the Interaction model when it is
serialised for Activity Stream: `notes` and `theme`.

This commit adds them to the serializer.

Co-Authored-By: Harriet H-W <16647486+Harriethw@users.noreply.github.com>
Co-Authored-By: Edmond DeLos Reyes <28296624+dredmonds@users.noreply.github.com>